requirements.txt
asgiref==3.8.1
attrs==23.2.0
backports.zoneinfo==0.2.1
coreschema==0.0.4
Django==4.2
djangorestframework==3.15.1
drf-yasg==1.21.7
importlib_resources==6.4.0
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
packaging==24.0
pkgutil_resolve_name==1.3.10
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
zipp==3.18.2
